requirements.txt
streamlit
pandas
plotly
sqlalchemy
psycopg2-binary
xgboost
scikit-learn
requests
beautifulsoup4
python-dotenv
tenacity
schedule
streamlit-extras